+++ bin/powowshell.psm1
@@ -50,18 +50,20 @@ function Invoke-PowowShell {
             $p1 = $p1.replace("!", "$Workspace\");
         }
     }
+    # Get back to the location of the caller
+    Pop-Location
 	ForEach ($Cmd in $Command) {
 		try {
-			Write-Verbose "`".\$Cmd.ps1`" $p1 $p2 $p3"
-			if ($p3) {& ".\$Cmd.ps1" $p1 $p2 $p3}
-			elseif ($p2) {& ".\$Cmd.ps1" $p1 $p2}
-			elseif ($p1) {& ".\$Cmd.ps1" $p1}
-			else {& ".\$Cmd.ps1"}
+			Write-Verbose "`"$BinPath\$Cmd.ps1`" $p1 $p2 $p3"
+			if ($p3) {& "$BinPath\$Cmd.ps1" $p1 $p2 $p3}
+			elseif ($p2) {& "$BinPath\$Cmd.ps1" $p1 $p2}
+			elseif ($p1) {& "$BinPath\$Cmd.ps1" $p1}
+			else {& "$BinPath\$Cmd.ps1"}
 		} catch {
 			#Write-Error "Error in '$cmd' command:" + $_
 			throw $_
 		} finally {
-            Pop-Location
+            
         }
 	}
 }