+++ ide/js/pipeline-manager.js
@@ -196,6 +196,25 @@ let pipelineManager = (function() {
             return null;
         },
         /**
+         * Return all inputs available to a step
+         *  i.e. the outputs of all steps previous to it
+         * @param {string} id 
+         * @returns {string[]} Step IDs of available outputs
+         */
+        getAvailableInputs: function(id) {
+            let res = [];
+            let col = parseCol(id);
+            let row = parseRow(id);
+            for (let c=1; c<=COLS; c++)
+                for (let r=1; r<=ROWS; r++)
+                    if (c<col || c==col && r < row) {
+                        let step = this.getStep(c, r);
+                        // TODO: Check step as outputs!
+                        if (step.reference) res.push(step.id);
+                    }
+            return res;
+        },
+        /**
          * Return the definition for unit testing
          * @returns {Object} The in-memory pipeline which was loaded
          */
@@ -339,6 +358,10 @@ if (typeof process !== "undefined") {
 (function pipelineManagerTest(verbose) {
     let tests = 0;
     let fails = 0;
+
+    let PM = pipelineManager;
+    let COLS = pipelineManager.pipeCols;
+
     function assert(cond, msg, exception) {
         if (verbose) console.log("\x1b[36m", "... " + msg);
         try {
@@ -387,58 +410,59 @@ if (typeof process !== "undefined") {
 
         console.clear();
 
-        let COLS = pipelineManager.pipeCols;
-
         // Check a new pipeline
-        pipelineManager.reset();
-        assert(pipelineManager.columnCount()===COLS.length, "We have the right number of columns ("+COLS.length+")")
-        assert(pipelineManager.rowCount()===9, "We have the right number of rows (9)")
-        assert(pipelineManager.nextRow("A")===1, "Next empty row in column A is 1")
+        PM.reset();
+        assert(PM.columnCount()===COLS.length, "We have the right number of columns ("+COLS.length+")")
+        assert(PM.rowCount()===9, "We have the right number of rows (9)")
+        assert(PM.nextRow("A")===1, "Next empty row in column A is 1")
         
         // Getting and Adding Steps
-        assert('pipelineManager.getStep(1, 1).reference === null', "Step A1 is initialized")
-        assert(()=>pipelineManager.addComponent(1, 1, {reference: "foo"}).reference === "foo", "Step A1 is set to foo")
-        assert(pipelineManager.nextRow("A")===2, "Next empty row in column A is 2")
-        assert('pipelineManager.addComponent("B", null, {reference: "bar"}).reference === "bar"', "Step B1 is set to bar")
-        assert('pipelineManager.addComponent("C3", null, testComponent).reference=="CSV2JSON.ps1"', "Add component to wrong row", true)
-        assert('pipelineManager.addComponent("C1", null, testComponent).reference=="CSV2JSON.ps1"', "Add C1 component")
-        assert('pipelineManager.addComponent("C", null, testComponent).reference=="CSV2JSON.ps1"', "Add C component")
-        assert('pipelineManager.addComponent("C", null, testComponent).reference=="CSV2JSON.ps1"', "Add another C component")
-        assert('pipelineManager.getStep("C", 3).reference !== null', "Step C3 is set")
-        let step = pipelineManager.getStep("C3");
-        assert(()=>pipelineManager.setStep(step), "Set step works");
+        assert('PM.getStep(1, 1).reference === null', "Step A1 is initialized")
+        assert(()=>PM.addComponent("1", 1, {reference: "foo"}).reference === "foo", "Step A1 is set to foo")
+        assert(PM.nextRow("A")===2, "Next empty row in column A is 2")
+        assert('PM.addComponent("B", null, {reference: "bar"}).reference === "bar"', "Step B1 is set to bar")
+        assert('PM.addComponent("C3", null, testComponent).reference=="CSV2JSON.ps1"', "Add component to wrong row", true)
+        assert('PM.addComponent("C1", null, testComponent).reference=="CSV2JSON.ps1"', "Add C1 component")
+        assert('PM.addComponent("C", null, testComponent).reference=="CSV2JSON.ps1"', "Add C component")
+        assert('PM.addComponent("C", null, testComponent).reference=="CSV2JSON.ps1"', "Add another C component")
+        assert('PM.getStep("C", 3).reference !== null', "Step C3 is set")
+        let step = PM.getStep("C3");
+        assert(()=>PM.setStep(step), "Set step works");
         assert(()=>step.parameters.length===3, "CSV2JSON should have 3 parameters");
 
+        // Available inputs
+        assert(()=>PM.getAvailableInputs("C3").length==4, "Step 3 has 2 available inputs");
+
         // Moving steps
-        assert(()=>pipelineManager.moveStep("C3", "A1"), "Step can't be moved over existing step", true);
-        assert(()=>pipelineManager.getStep("D4").reference===null, "Step D4 should be empty");
-        assert(()=>pipelineManager.getStep("A1").reference!==null, "Step A1 should not be empty");
-        assert(()=>pipelineManager.getStep("C3").reference!==null, "Step C3 should not be empty"); 
-        pipelineManager.moveStep("C3", "D4")
-        let c3 = pipelineManager.getStep("C3");
-        let d4 = pipelineManager.getStep("D4");
+        assert(()=>PM.moveStep("C3", "A1"), "Step can't be moved over existing step", true);
+        assert(()=>PM.getStep("D4").reference===null, "Step D4 should be empty");
+        assert(()=>PM.getStep("A1").reference!==null, "Step A1 should not be empty");
+        assert(()=>PM.getStep("C3").reference!==null, "Step C3 should not be empty"); 
+        PM.moveStep("C3", "D4")
+        let c3 = PM.getStep("C3");
+        let d4 = PM.getStep("D4");
         assert(()=>d4.reference=="CSV2JSON.ps1" && c3.reference === null, "Step C3 moved to D4");
 
         // Removing steps
-        pipelineManager.removeStep("D4");
-        assert(()=>pipelineManager.getStep("D4").reference === null, "Step D4 is removed")
+        PM.removeStep("D4");
+        assert(()=>PM.getStep("D4").reference === null, "Step D4 is removed")
 
         // Import
-        assert(()=>pipelineManager.import(testPipeline), "Pipeline import");
-        assert(()=>pipelineManager.getStep("A1").reference.match(/File/), "Test pipeline step B1 is File")
-        assert(()=>pipelineManager.getStep("B1").reference.match(/CSV2JSON/), "Test pipeline step B1 is CSV2JSON")
-        assert(()=>pipelineManager.getStep("C1").reference.match(/SelectFields/), "Test pipeline step C1 is SelectFields")
-        assert(()=>typeof pipelineManager.getStep("A1").parameters.Path === "string", "Step A1 of imported pipeline has a 'Path' parameter")
+        assert(()=>PM.import(testPipeline), "Pipeline import");
+        assert(()=>PM.getStep("A1").reference.match(/File/), "Test pipeline step B1 is File")
+        assert(()=>PM.getStep("B1").reference.match(/CSV2JSON/), "Test pipeline step B1 is CSV2JSON")
+        assert(()=>PM.getStep("C1").reference.match(/SelectFields/), "Test pipeline step C1 is SelectFields")
+        assert(()=>typeof PM.getStep("A1").parameters.Path === "string", "Step A1 of imported pipeline has a 'Path' parameter")
         
         // Export
-        let myPipeline = pipelineManager.export();
+        let myPipeline = PM.export();
         assert(()=>myPipeline.steps.length===3, "Pipeline export has 3 steps")
-        let pipelineDef = pipelineManager.getDefinition();
+        let pipelineDef = PM.getDefinition();
         assert(()=>myPipeline.id===pipelineDef.id, "Export has same id")
         let pipelineStr = JSON.stringify(myPipeline);
         //fs.writeFileSync("C:\\temp\\pipeline.json", pipelineStr, "utf-8");
 
-        //console.log(pipelineManager.getStep("C1"))
+        //console.log(PM.getStep("C1"))
         //columnsToString();
 
     } catch(e) {