+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 CHANGELOG
 
-v0.1.1 16.03.2018
+v0.1.3 23.03.2019
+* BUGFIX: Paths broken in pow build/run/verify etc
+* Dynamically load available inputs selector on step
+* IDE Forms
+* Input/Output Descriptions of components
+
+v0.1.1 16.03.2019
 * Drop DOS support
 * Generate components.json with `pow components ./examples/components export | ConvertTo-JSON`